+++ iotoolkit/mat_data.py
@@ -81,7 +81,7 @@ class MatData:
     @staticmethod
     def isscalar(data):
         type_str = type(data)
-        return type_str in [np.uint8,np.int8,np.uint16,np.int16,np.int32,np.uint32,np.int64,np.uint64,np.float,np.float64,np.str_,np.str]
+        return type_str in [np.uint8,np.int8,np.uint16,np.int16,np.int32,np.uint32,np.int64,np.uint64,np.float32,np.float64,np.str_,np.str]
 
     def item(self):
         return self.data
